--- docs/reference/contrib/repository/abc.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-abc
-===
-
-This page has moved to :doc:`/reference/repository/abc`
