--- docs/reference/contrib/repository/exceptions.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-exceptions
-==========
-
-This page has moved to :doc:`/reference/repository/exceptions`
