--- docs/reference/contrib/repository/filters.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-filters
-=======
-
-This page has moved to :doc:`/reference/repository/filters`
