--- docs/reference/contrib/repository/handlers.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-handlers
-=========
-
-This page has moved to :doc:`/reference/repository/handlers`
