--- docs/reference/contrib/repository/testing.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-testing
-=======
-
-This page has moved to :doc:`/reference/repository/testing`
